+++ Sources/OnboardingKit/Onboarding.swift
@@ -20,55 +20,34 @@ import Foundation
  different rules to provide a different experience.
  
  Using an onboarding is easy. Just create an instance of the
- onboarding, and call ``tryPresent(delay:action:)`` when you
- want it to be presented. It will inspect its internal rules
- and state to determine if it should be presented of not.
+ specific onboarding type that you want to use, then trigger ``tryPresent(after:action:)`` to attempt to present it. The
+ onboarding will inspect its internal rules and only perform
+ the presentation action if needed.
  
  Call ``reset()`` to reset the onboarding state, at any time.
  */
 open class Onboarding: Identifiable {
     
-    /// Create an onboarding experience.
-    ///
-    /// - Parameters:
-    ///   - id: The unique onboarding id.
-    ///   - defaults: The `UserDefaults` instance to use, by default `.standard`.
+    /// Create a standard onboarding.
     public init(
         id: String,
-        defaults: UserDefaults = .standard
+        store: UserDefaults = .standard
     ) {
         self.id = id
-        self.defaults = defaults
+        self.store = store
     }
-    
-    public typealias UserId = String
 
 
     /// The unique onboarding id.
     public let id: String
 
     /// The `UserDefaults` instance to use.
-    public let defaults: UserDefaults
-
-
-    /// Whether or not the onboarding has been presented.
-    open var hasBeenPresented: Bool {
-        presentationCount > 0
-    }
+    public let store: UserDefaults
+    
     
     /// Whether or not the onboarding should be presented.
     open var shouldBePresented: Bool {
         presentationCount == 0
-    }
-    
-    
-    /// Register an onboarding presentation.
-    ///
-    /// The function can be called to manually set the state
-    /// of the onboarding, e.g. if a user has indicated that
-    /// she/he knows about what it's intended to communicate.
-    open func registerPresentation() {
-        presentationCount += 1
     }
     
     /// Reset the presentation state of the onboarding.
@@ -76,19 +55,31 @@ open class Onboarding: Identifiable {
         presentationCount = 0
     }
     
-    /// Try to present the onboarding with a certain action.
+    /// Try to present the onboarding with a provided action,
+    /// after an optional delay.
     ///
-    /// The presentation will be ignored if the onboarding's
-    /// behavior or state indicates that it does not have to
-    /// be presented.
+    /// The presentation will be ignored when the onboarding
+    /// does not have to be presented.
     open func tryPresent(
         after seconds: TimeInterval = 0,
         action: @escaping () -> Void
     ) {
         guard shouldBePresented else { return }
-        let queue = DispatchQueue.main
-        queue.asyncAfter(deadline: .now() + seconds, execute: action)
-        registerPresentation()
+        present(after: seconds, action: action)
+    }
+    
+    /// Present the onboarding with the provided action.
+    open func present(
+        after seconds: TimeInterval = 0,
+        action: @escaping () -> Void
+    ) {
+        presentationCount += 1
+        if seconds > 0 {
+            let queue = DispatchQueue.main
+            queue.asyncAfter(deadline: .now() + seconds, execute: action)
+        } else {
+            action()
+        }
     }
 }
 
@@ -96,8 +87,8 @@ public extension Onboarding {
 
     /// The number of times an onboarding has been presented.
     var presentationCount: Int {
-        get { defaults.integer(forKey: presentationCountKey) }
-        set { defaults.set(newValue, forKey: presentationCountKey) }
+        get { store.integer(forKey: presentationCountKey) }
+        set { store.set(newValue, forKey: presentationCountKey) }
     }
     
     /// The unique storage key to use for persisting values.