+++ Tests/OnboardingKitTests/Onboardings/ConditionalOnboardingTests.swift
@@ -11,26 +11,59 @@ import XCTest
 
 final class ConditionalOnboardingTests: XCTestCase {
 
-    var onboarding: Onboarding!
+    var onboarding: ConditionalOnboarding!
+    
+    var condition = false
+    
+    override func setUp() {
+        onboarding = .init(
+            id: "onboarding",
+            condition: { self.condition }
+        )
+    }
 
     override func tearDown() {
         onboarding.reset()
         XCTAssertEqual(onboarding.presentationCount, 0)
     }
-
-    func test_shouldBePresented_isTrueIfConditionIsTrueAndOnboardingHasNotBeenPresented() {
-        onboarding = ConditionalOnboarding(id: "id", condition: { true })
-        XCTAssertTrue(onboarding.shouldBePresented)
+    
+    
+    func testInitializerSetsUpOnboarding() {
+        onboarding = .init(
+            id: "test",
+            store: TestUserDefaults(),
+            condition: { self.condition }
+        )
+        XCTAssertEqual(onboarding.id, "test")
+        XCTAssertTrue(onboarding.store is TestUserDefaults)
+        XCTAssertFalse(onboarding.condition())
+        condition = true
+        XCTAssertTrue(onboarding.condition())
     }
-
-    func test_shouldBePresented_isFalseIfConditionIsTrueButOnboardingHasBeenPresented() {
-        onboarding = ConditionalOnboarding(id: "id", condition: { true })
-        onboarding.registerPresentation()
-        XCTAssertFalse(onboarding.shouldBePresented)
+    
+    func testTryPresentOnlyWorksOnceIfConditionIsTrue() {
+        var counter = 0
+        onboarding.tryPresent { counter += 1 }
+        XCTAssertEqual(counter, 0)
+        condition = true
+        onboarding.tryPresent { counter += 1 }
+        XCTAssertEqual(counter, 1)
+        onboarding.tryPresent { counter += 1 }
+        XCTAssertEqual(counter, 1)
     }
-
-    func test_shouldBePresented_isFalseIfConditionIsFalse() {
-        onboarding = ConditionalOnboarding(id: "id", condition: { false })
-        XCTAssertFalse(onboarding.shouldBePresented)
+    
+    func testPresentWorksEveryTime() {
+        var counter = 0
+        onboarding.present { counter += 1 }
+        XCTAssertEqual(counter, 1)
+        onboarding.present { counter += 1 }
+        XCTAssertEqual(counter, 2)
+    }
+    
+    func testResetResetsOnboardingState() {
+        onboarding.present {}
+        XCTAssertEqual(onboarding.presentationCount, 1)
+        onboarding.reset()
+        XCTAssertEqual(onboarding.presentationCount, 0)
     }
 }