+++ Tests/OnboardingKitTests/TestUtils.swift
@@ -1,0 +1,11 @@
+//
+//  TestUtils.swift
+//  OnboardingKitTests
+//
+//  Created by Daniel Saidi on 2024-03-10.
+//  Copyright © 2024 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+class TestUserDefaults: UserDefaults {}