+++ RELEASE_NOTES.md
@@ -13,8 +13,11 @@ Older release notes up until 6.0 have been removed.
 ### 💥 Breaking changes
 
 * All deprecated code has been removed.
+
 * `Hint` has been removed.
-* `Onboarding` no longer has a `userId` since it's not needed.
+* `Onboarding` no longer has a `userId` property.
+* `Onboarding` no longer has a `hasBeenPresented` property.
+* `Onboarding` no longer has a `registerPresentation` function.